+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New features
 
+* [#4252](https://github.com/bbatsov/rubocop/issues/4252): Add new `Style/TrailingBodyOnMethodDefinition` cop. ([@garettarrowood][])
 * [#4650](https://github.com/bbatsov/rubocop/issues/4650): Add new `Style/StringHashKeys` cop. ([@donjar][])
 * [#1583](https://github.com/bbatsov/rubocop/issues/1583): Add a quiet formatter. ([@drenmi][])
 * Add new `Style/RandomWithOffset` cop. ([@donjar][])
@@ -3022,3 +3023,4 @@
 [@michniewicz]: https://github.com/michniewicz
 [@bgeuken]: https://github.com/bgeuken
 [@mtsmfm]: https://github.com/mtsmfm
+[@garettarrowood]: https://github.com/garettarrowood